refactor(#3276): decompose compilePropertyAccess slice 2 — extract 4 more guard bands
Wave B (subtask of #3182), stacked on slice 1. Extract 4 more cohesive
early-return guard bands from compilePropertyAccess into
property-access-dispatch.ts: tryPrivateIdentifierRead, trySuperAndImportMetaRead,
tryGlobalThisAndProcessRead, tryIdentifierNamespaceAndStaticReceiverRead.

property-access.ts: 7107 -> 6473 LOC (-634; cumulative 7989 -> 6473 = -1516).
Byte-identity: IDENTICAL 39/39 (gc/standalone/wasi). tsc 0. Smoke 11/11.

✓
